Add CMakeLists.txt prototype
We want to migrate from QMake to CMake in the future. This CMake file
allows for building the project without errors but our cuboid does not
move. This is probably due to the fact that we have not added shaders
files to the project yet.

This helps: https://doc.qt.io/qt-5/cmake-command-reference.html